+++ src/lib.rs
@@ -2,8 +2,10 @@ mod cli;
 mod error;
 mod letters_boxed;
 mod prepare_words;
+mod shape;
 
 pub use cli::{Cli, Commands};
 pub use error::Error;
 pub use letters_boxed::{LettersBoxed, Shuffle};
 pub use prepare_words::PrepareWords;
+pub use shape::Shape;