+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- ✨ add generate command(pr [#43])
+
 ## [0.1.11] - 2025-04-02
 
 ### Added
@@ -155,6 +161,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#40]: https://github.com/jerus-org/slb/pull/40
 [#41]: https://github.com/jerus-org/slb/pull/41
 [#42]: https://github.com/jerus-org/slb/pull/42
+[#43]: https://github.com/jerus-org/slb/pull/43
+[Unreleased]: https://github.com/jerus-org/slb/compare/v0.1.11...HEAD
 [0.1.11]: https://github.com/jerus-org/slb/compare/v0.1.10...v0.1.11
 [0.1.10]: https://github.com/jerus-org/slb/compare/v0.1.9...v0.1.10
 [0.1.9]: https://github.com/jerus-org/slb/compare/v0.1.8...v0.1.9