+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ✨ add solution module for word chain puzzle(pr [#47])
 - ✨ add progress bar to solution generation(pr [#48])
 
+### Changed
+
+- ♻️ refactor(cli)-streamline command structure and logic(pr [#49])
+
 ## [0.1.12] - 2025-04-03
 
 ### Added
@@ -177,6 +181,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#46]: https://github.com/jerus-org/slb/pull/46
 [#47]: https://github.com/jerus-org/slb/pull/47
 [#48]: https://github.com/jerus-org/slb/pull/48
+[#49]: https://github.com/jerus-org/slb/pull/49
 [Unreleased]: https://github.com/jerus-org/slb/compare/v0.1.12...HEAD
 [0.1.12]: https://github.com/jerus-org/slb/compare/v0.1.11...v0.1.12
 [0.1.11]: https://github.com/jerus-org/slb/compare/v0.1.10...v0.1.11