+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - ♻️ refactor(cli)-streamline command structure and logic(pr [#49])
+- ♻️ refactor(main)-update default source file name(pr [#50])
 
 ## [0.1.12] - 2025-04-03
 
@@ -182,6 +183,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#47]: https://github.com/jerus-org/slb/pull/47
 [#48]: https://github.com/jerus-org/slb/pull/48
 [#49]: https://github.com/jerus-org/slb/pull/49
+[#50]: https://github.com/jerus-org/slb/pull/50
 [Unreleased]: https://github.com/jerus-org/slb/compare/v0.1.12...HEAD
 [0.1.12]: https://github.com/jerus-org/slb/compare/v0.1.11...v0.1.12
 [0.1.11]: https://github.com/jerus-org/slb/compare/v0.1.10...v0.1.11