+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.13] - 2025-04-05
 
 ### Added
 
@@ -186,7 +186,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#49]: https://github.com/jerus-org/slb/pull/49
 [#50]: https://github.com/jerus-org/slb/pull/50
 [#51]: https://github.com/jerus-org/slb/pull/51
-[Unreleased]: https://github.com/jerus-org/slb/compare/v0.1.12...HEAD
+[0.1.13]: https://github.com/jerus-org/slb/compare/v0.1.12...v0.1.13
 [0.1.12]: https://github.com/jerus-org/slb/compare/v0.1.11...v0.1.12
 [0.1.11]: https://github.com/jerus-org/slb/compare/v0.1.10...v0.1.11
 [0.1.10]: https://github.com/jerus-org/slb/compare/v0.1.9...v0.1.10